requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.11
filetype==1.2.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
gritql==0.1.5
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.20.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-experimental==0.3.4
langchain-google-genai==2.0.11
langchain-openai==0.3.7
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langgraph==0.3.2
langgraph-api==0.0.27
langgraph-checkpoint==2.0.16
langgraph-cli==0.1.74
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.3.11
lark==1.2.2
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.3
openai==1.65.2
orjson==3.10.15
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.13.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.0
structlog==23.3.0
tenacity==9.0.0
tiktoken==0.9.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
widgetsnbextension==4.0.13
yarl==1.18.3
zstandard==0.23.0
